apps/api/pyproject.toml
[project]
name = "devprod-api"
version = "0.1.0"
description = "DevProd API service"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.1,<0.117",
  "pydantic-settings>=2.10.1,<3",
  "uvicorn>=0.35.0,<0.36",
  "jsonschema>=4.25.1,<5"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<0.29",
  "pytest>=8.4.2,<9",
  "pytest-asyncio>=1.2.0,<2",
  "ruff>=0.13.1,<0.14",
  "mypy>=1.18.2,<2"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-p no:cacheprovider"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
plugins = []

[[tool.mypy.overrides]]
module = ["jsonschema", "jsonschema.*"]
ignore_missing_imports = true

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
